+++ src/lib/codes.ts
@@ -9,6 +9,7 @@
  */
 
 export const TRANSACTION_NOT_FOUND = 10;
+export const ERROR_ARANGO_CONFLICT = 1200;
 export const ANALYZER_NOT_FOUND = 1202;
 export const DOCUMENT_NOT_FOUND = 1202;
 export const COLLECTION_NOT_FOUND = 1203;